Fixed an issue with local and particularly global namespacing
The issue was only shown because of the recent addition of list support
in the dict merger. The issue was that the global_updates dict was not
being cleared before the final recursive update on the global namespace.
The result was the that the recursive dict was being run twice and for
lists, this meant the update happened twice, for lists yielding double
entries.

The fix was to move the recursive update inside the post-hook conditional
which means that the global updates are only updated for a post-hook
if a post-hook actually occured.

This change was also made for the pre-hook to make sure they worked in
the same way.