Fix deadlock happening in CFME integration_tests
When the artifactor process finishes in the CFME integration tests,
there is nothing listening on the tcp. But some logging related plugin
in the CFME integration_tests is causing riggerlib trying to connect
even after the artifactor process is stopped and it is sending a probe
message. Thus the probe request is in the outgoing queue causing a hang
in GC time when the Python is about to exit. ZMQ does not want to
teardown the context of the socket in that case...

There is probably a bug in pyzmq as
http://zguide.zeromq.org/page:all says:

```
First, do not try to use the same socket from multiple threads. Please
don't explain why you think this would be excellent fun, just please
don't do it. Next, you need to shut down each socket that has ongoing
requests. The proper way is to set a low LINGER value (1 second), and
then close the socket. If your language binding doesn't do this for you
automatically when you destroy a context, I'd suggest sending a patch.
```

There is a code that should handle this situation in the riggerlib but
I think it was not in effect due to the line

    socket = self.__dict__.pop('_socket')

caused an KeyError as the _socket was not set because the
`poll.poll()` called in `_very_lazy_pirate_request` in `_ensure_connected`
returned empty sequence, causing KeyError there.

The errors from the riggerlib got silenced in the `fire_hook` method with the

    except Exception:
        return None

which does smell quite anti-patterny to me and perhaps should be improved in
the future.